app/requirements/base.txt
dj-database-url==0.5.0
Django==4.0.6
django-cors-headers==3.13.0
django-filter==22.1
psycopg2-binary==2.9.3
python-decouple==3.6
celery==5.2.7
flower==1.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
Faker==13.15.0
pillow==9.5.0
pyotp==2.8.0 
twilio===8.1.0